Move all UI objects into their own folder
