+++ FraterniTree/UserInterface/AboutCompanyDialog.resx
