+++ FraterniTree/UserInterface/EditBrotherWithoutBigDialog.resx
